Extract pure buildFileExportRecord into self-contained module
The submissions-export-files unit test imported buildFileExportRecord from
submissionFilesExportStream.ts, which transitively pulls Deno-only modules
(supabase client, .ts-extension imports) into the Next.js tsc program and
broke the type-check. Move the pure helper into a self-contained
submissionFileExportRecord.ts (matching the repo's testable-CLI-module
convention), re-export it for the Deno consumer, and point the test there.
